v0.35: comprehensive quality audit fixes — security, performance, correctness
Concurrency:
- Fix guaranteed deadlock in leave_document (tokio::RwLock double-acquire)
- Standardize lock ordering: rooms before clients in ConnectionManager

Security:
- Fix CORS to respect configured allowed_origins (was always Any)
- Fix disabled CORS being MORE permissive than enabled
- Fix 5 webhook delivery sites: let _ = future → tokio::spawn
- Fix SQL injection in SSG: format! interpolation → bind params
- Enforce JWT secret validation at startup (Err, not warning)
- Unify auth error responses (all 401, no information leakage)
- Add password strength validation (8+ chars, upper, lower, digit)
- Narrow auth bypass from blanket prefix to explicit route list

Performance:
- Eliminate N+1 queries in 4 endpoints (organization, space list variants)
- Add count_members_batch/count_documents_batch using ANY(::uuid[])
- Replace blocking std::fs with tokio::fs in async context (sync.rs, main.rs)

Determinism:
- Replace HashMap with BTreeMap in 17+ serialized structs for stable JSON
- Sort list_repositories, search_files, collect_markdown_files results

WASM Safety:
- Remove 2 unsafe blocks (Uint8Array::view → safe copy)
- Replace Arc<Mutex<ApiClient>> with Rc<RefCell<ApiClient>> in WASM

API Hygiene:
- Add pub(crate) to 20+ internal items across 6 crates
- Add crate-level doc comments to 6 crates
- Remove 8 #[allow(unused_imports)] from frontend
- Declare staging feature in 7 crate Cargo.toml files

Clippy: 100+ auto-fixes (needless borrows, redundant closures, unsigned_abs, etc.)
All 794 tests pass, 0 clippy errors with -D warnings

+++ tachyon/crates/search/src/query.rs
@@ -63,7 +63,7 @@ impl QueryEngine {
         let query = self.parse_query(request, schema)?;
 
         // Create collector with pagination
-        let limit = request.page_size as usize;
+        let limit = request.page_size;
         let _offset = request.offset();
 
         let collector = TopDocs::with_limit(limit);
